0.2.0 commercial readiness push
Phase 1 — Bug Fixes:
- Codegen: post-CSE constant propagation eliminates trivial temps (1.0, 0.0^2)
- Codegen: Add detects Neg children, emits subtraction not '+ (-x)'
- Codegen: fractions as decimal literals (-0.25_f64 not -1_f64/4_f64)
- Codegen: reduced excessive parenthesization
- Codegen: hero output 1448 → 878 bytes for 3-DOF Jacobian
- Robotics: fk_position calls eval() to eliminate cos(0)/sin(0) artifacts
- Dynamics: all outputs call eval() to simplify constant subexpressions
- Solve: trig inversion fast path (sin(x)=1/2 → π/6)
- Polysys: symbolic fallback for irrational Gröbner roots (x²+y²=3,x=y)
- Macro: expr!(1/2) now emits rational(1,2), expr!(-1/2) fixed
- LaTeX: sin²(x) renders as conventional \sin^{2}(x)
- Clippy: zero warnings on all targets (was 60+ in lib, 20+ in tests)

Phase 2 — Tutorial (4,051 lines across 9 pages):
- 01-introduction: what is symbolic math, why symplex
- 02-getting-started: installation, expr!, subs, eval, compile
- 03-gotchas: 8 common pitfalls
- 04-simplification: full toolkit (536 lines)
- 05-calculus: diff, integrate, limits, series, Laplace (601 lines)
- 06-solving: polynomial through Gröbner, ODE (605 lines)
- 07-matrices: construction through codegen (576 lines)
- 08-code-generation: THE KILLER PAGE, end-to-end pipeline (866 lines)

Phase 3-4 — Examples and Docs:
- Rewrote 10 examples (1,105 → 3,514 total lines, 3.2x increase)
- Added 3 new examples: ode_solving, complex_numbers, laplace_transforms
- Added docs/sympy-comparison.md (425 lines)
- Updated IMPLEMENTATION_PLAN: test timing docs, resolved 6 known limitations

Stats: 4,560 tests passing, 0 failing, 0 clippy warnings
Known Limitations resolved: #2, #17, #18, #19, #20, #23

+++ benches/canonicalization.rs
@@ -47,7 +47,7 @@ fn bench_diff(c: &mut Criterion) {
     });
 
     c.bench_function("diff_polynomial_deg10", |b| {
-        let terms: Vec<Ex> = (0..=10).map(|n| &x.powi(n) * (n + 1) as i64).collect();
+        let terms: Vec<Ex> = (0..=10).map(|n| &x.powi(n) * (n + 1)).collect();
         let poly = Ex::sum_of(&ctx, terms);
         b.iter(|| poly.diff(&x));
     });
@@ -544,7 +544,7 @@ fn bench_codegen(c: &mut Criterion) {
     let x = ctx.symbol("x");
 
     c.bench_function("codegen_poly_deg10", |b| {
-        let terms: Vec<Ex> = (0..=10).map(|n| &x.powi(n) * (n + 1) as i64).collect();
+        let terms: Vec<Ex> = (0..=10).map(|n| &x.powi(n) * (n + 1)).collect();
         let poly = Ex::sum_of(&ctx, terms);
         b.iter(|| black_box(&poly).to_rust_fn("f", &["x"]))
     });
@@ -555,7 +555,7 @@ fn bench_codegen(c: &mut Criterion) {
     });
 
     c.bench_function("lambdify_poly_deg10", |b| {
-        let terms: Vec<Ex> = (0..=10).map(|n| &x.powi(n) * (n + 1) as i64).collect();
+        let terms: Vec<Ex> = (0..=10).map(|n| &x.powi(n) * (n + 1)).collect();
         let poly = Ex::sum_of(&ctx, terms);
         b.iter(|| black_box(&poly).compile(&["x"]))
     });
